update for robot replay

+++ Device.py
@@ -58,7 +58,7 @@ class Device:
             if matched_element is not None:
                 self.execute_action('click', [(int(matched_element.center_x / self.detect_resize_ratio), int(matched_element.center_y / self.detect_resize_ratio))])
             else:
-                coord = (int(action['coordinate'][0][0] / screen_ratio), action['coordinate'][0][1] / screen_ratio)
+                coord = (int(action['coordinate'][0][0] / screen_ratio), int(action['coordinate'][0][1] / screen_ratio))
                 self.execute_action('click', [coord])
         elif action['type'] == 'swipe':
             start_coord = (int(action['coordinate'][0][0] / screen_ratio), action['coordinate'][0][1] / screen_ratio)
